requirements.txt
aniso8601==3.0.0
certifi==2018.1.18
chardet==3.0.4
Django==1.11.10
django-facebook==6.0.3
django-geoposition==0.3.1
django-material==1.1.0
django-simple-history==1.9.1
easy-thumbnails==2.5
graphene==2.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
oauthlib==2.0.7
Pillow==5.0.0
promise==2.1
PyJWT==1.6.1
python-openid==2.2.5
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typing==3.6.4
Unidecode==1.0.22
urllib3==1.22
